Special case `&str` if used in type ascription within 'construct_with' attribute.
Usually, the requirement is that `std::str::FromStr` is implemented for the receiver type. However, this doesn't work with `&str`:
`std::str::FromStr` is not implemented for `&str` (presumably due to lifetime issues and lack of GATs). However, we already get a `&str` from the regex, parsing to `&str` is unnecessary.
Therefore, if the receiver type is ascribed as `&str`, we directly use the `&str` we get from the regex.